+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/client/ByProjectKeyInStoreKeyByStoreKeyMeGet.java
@@ -20,7 +20,13 @@ import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
 
 /**
- *
+ *  <p>Returns a Customer for a given Query Predicate in a Store. Returns a <code>200 OK</code> status if successful.</p>
+ *  <p>A ResourceNotFound error is returned in the following scenarios:</p>
+ *  <ul>
+ *   <li>If no Customer exists in the Store for the given Query Predicate.</li>
+ *   <li>If a Customer exists in the Store for the given Query Predicate, but does not have an <code>id</code> value that matches the customer:{id} scope.</li>
+ *   <li>If a Customer exists for the given Query Predicate but is associated with a different Store than what is specified in the <code>manage_my_profile:{projectKey}:{storeKey}</code> scope.</li>
+ *  </ul>
  *
  * <hr>
  * <div class=code-example>