+++ commercetools/commercetools-sdk-java-api/src/test/java-generated/com/commercetools/api/client/resource/ByProjectKeyMeBusinessUnitsKeyByKeyTest.java
@@ -99,31 +99,7 @@ public class ByProjectKeyMeBusinessUnitsKeyByKeyTest {
                         .businessUnits()
                         .withKey("test_key")
                         .post(com.commercetools.api.models.me.MyBusinessUnitUpdate.of())
-                        .createHttpRequest(), "post", "test_projectKey/me/business-units/key=test_key", },
-                new Object[] {
-                        apiRoot.withProjectKey("test_projectKey")
-                                .me()
-                                .businessUnits()
-                                .withKey("test_key")
-                                .delete()
-                                .withVersion(2)
-                                .createHttpRequest(),
-                        "delete", "test_projectKey/me/business-units/key=test_key?version=2", },
-                new Object[] {
-                        apiRoot.withProjectKey("test_projectKey")
-                                .me()
-                                .businessUnits()
-                                .withKey("test_key")
-                                .delete()
-                                .withExpand("expand")
-                                .createHttpRequest(),
-                        "delete", "test_projectKey/me/business-units/key=test_key?expand=expand", },
-                new Object[] { apiRoot.withProjectKey("test_projectKey")
-                        .me()
-                        .businessUnits()
-                        .withKey("test_key")
-                        .delete()
-                        .createHttpRequest(), "delete", "test_projectKey/me/business-units/key=test_key", } };
+                        .createHttpRequest(), "post", "test_projectKey/me/business-units/key=test_key", } };
     }
 
     @DataProvider
@@ -149,23 +125,6 @@ public class ByProjectKeyMeBusinessUnitsKeyByKeyTest {
                         .me()
                         .businessUnits()
                         .withKey("test_key")
-                        .post(com.commercetools.api.models.me.MyBusinessUnitUpdate.of()), },
-                new Object[] { apiRoot.withProjectKey("test_projectKey")
-                        .me()
-                        .businessUnits()
-                        .withKey("test_key")
-                        .delete()
-                        .withVersion(2), },
-                new Object[] { apiRoot.withProjectKey("test_projectKey")
-                        .me()
-                        .businessUnits()
-                        .withKey("test_key")
-                        .delete()
-                        .withExpand("expand"), },
-                new Object[] { apiRoot.withProjectKey("test_projectKey")
-                        .me()
-                        .businessUnits()
-                        .withKey("test_key")
-                        .delete(), } };
+                        .post(com.commercetools.api.models.me.MyBusinessUnitUpdate.of()), } };
     }
 }